% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffy.R
\docType{package}
\name{buffy}
\alias{buffy}
\alias{buffy-package}
\title{buffy: An R package with advanced buffering for geometry objects}
\description{
Provides functions for generating advanced types of buffers around geometry
objects.  Generates buffers on the WSG84 ellipsoid or in Euclidean geometry,
as appropriate.
}
\section{buffy functions}{

\code{\link{rectBuff}}
\code{\link{surfBuff}}
\code{\link{rad2deg}}
\code{\link{deg2rad}}
\code{\link{snap}}
}

