% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rad2deg.R
\name{rad2deg}
\alias{rad2deg}
\alias{deg2rad}
\title{Radians to degrees}
\usage{
rad2deg(x)

deg2rad(x)
}
\arguments{
\item{x}{numeric radians or degrees}
}
\value{
numeric degrees or radians
}
\description{
Convert radians to degrees
}
\section{Functions}{
\itemize{
\item \code{deg2rad}: Convert degrees to radians
}}

