% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectBuff.R
\name{rectBuff}
\alias{rectBuff}
\title{Construct rectangular buffers}
\usage{
rectBuff(x, d1, d2, b = 0)
}
\arguments{
\item{x}{point(s) of longitude-latitude in a row vector or matrix of two
columns, or an object of class SpatialPoints*}

\item{d1}{half length of bearing aligned distance(s) of class units}

\item{d2}{half length of bearing orthogonal distances(s) of class units}

\item{b}{bearing(s) of \code{d1} in degrees}
}
\value{
A SpatialPolygons object
}
\description{
Construct rectangular buffers around points of \code{x} with side-lengths
twice the values of \code{d1} and \code{d2}.  The axis defined by \code{d1}
will be aligned parallel with bearing \code{b}.  If \code{b} is missing,
then it is assumed to be 0 degrees.  If parameter \code{d2} is missing, then
this function constructs a square with length \code{d1}.
}
\examples{
p11 <- c(-92.44744828628, 34.566107548536)
p12 <- p11 + rnorm(2, sd=0.01)
p21 <- p12 + rnorm(2, sd=0.01)
p22 <- p21 + rnorm(2, sd=0.01)
b1 <- geosphere::bearing(p11, p12)
b2 <- geosphere::bearing(p21, p22)
sp::plot(rectBuff(p11, units::ud_units $mi, 2*units::ud_units $mi, b1))
sp::plot(with(units::ud_units, rectBuff(rbind(p11, p21), mi, 2*mi,
                                        c(b1, b2))))
}
\seealso{
\code{\link[geosphere]{bearing}}
}
