% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap.R
\name{snap}
\alias{snap}
\title{Snap polygons}
\usage{
snap(x, d)
}
\arguments{
\item{x}{an sf, sfc, or sfg}

\item{d}{an object of class units}
}
\value{
a vector of group identifiers
}
\description{
Group polygons together that are within some distance of each other
}
\details{
Code based on:
<https://gis.stackexchange.com/questions/254519/group-and-union-polygons-that-share-a-border-in-r>
}
